refactor(whispering): drop dead Rust-error remap, rename to local-preflight
The local-transcription.ts file carried five Rust-mirror error variants
(ModelLoadError, GpuError, AudioReadError, TranscriptionError,
UnexpectedLocalError) and a mapLocalTranscriptionError translation
that reconstructed them from commands.transcribeRecording's native
TranscriptionError. The map was identity for the four known variants
and a defensive wrapper for the catch-all.

A repo-wide audit confirmed mapLocalTranscriptionError was the only
error.name switch in the entire src tree. Nothing pattern-matched on
its outputs. The whole translation layer was plumbing, doubling the
maintenance surface against schema drift from Rust.

Drop the five mirror variants and the map. Pass Rust's TranscriptionError
through to the dispatcher's AnyTaggedError return type unchanged.
The transcribeRecording wrapper had one caller, so call
commands.transcribeRecording directly from the dispatcher's three
engine cases.

Rename the file to local-preflight.ts and the type to
LocalPreflightError so the name matches what is left: FE-side path
validation and engine-specific preflight guards
(CorruptedModelFile, InvalidMoonshineDirectoryName).

local-preflight.ts: 95 lines, was 159 in local-transcription.ts.
